requirements.txt
django
Pillow
django-localflavor
django-crispy-forms
crispy-bootstrap5
django-filter